Move empty state components to console shared package
