+++ frontend/packages/console-shared/locales/en/console-shared.json
@@ -119,6 +119,11 @@
   "View property descriptions": "View property descriptions",
   "Save": "Save",
   "View shortcuts": "View shortcuts",
+  "Restricted access": "Restricted access",
+  "You don't have access to this section due to cluster policy": "You don't have access to this section due to cluster policy",
+  "Error details": "Error details",
+  "No {{label}} found": "No {{label}} found",
+  "Not found": "Not found",
   "Extension error": "Extension error",
   "Show details": "Show details",
   "Oh no! Something went wrong.": "Oh no! Something went wrong.",
