+++ frontend/packages/console-shared/src/components/empty-state/index.ts
@@ -1,0 +1,3 @@
+export * from './AccessDenied';
+export * from './ConsoleEmptyState';
+export * from './EmptyBox';
