+++ frontend/public/components/utils/status-box/index.ts
@@ -1,6 +1,4 @@
-export * from './AccessDenied';
-export * from './EmptyBox';
-export * from './ConsoleEmptyState';
 export * from './StatusBox';
 
 export * from '@console/shared/src/components/loading';
+export * from '@console/shared/src/components/empty-state';
