+++ frontend/public/locales/en/public.json
@@ -1779,11 +1779,6 @@
   "For continued support, upgrade your cluster or transfer cluster ownership to an account with an active subscription.": "For continued support, upgrade your cluster or transfer cluster ownership to an account with an active subscription.",
   "Create new option \"{{option}}\"": "Create new option \"{{option}}\"",
   "Filter options": "Filter options",
-  "Restricted access": "Restricted access",
-  "You don't have access to this section due to cluster policy": "You don't have access to this section due to cluster policy",
-  "Error details": "Error details",
-  "No {{label}} found": "No {{label}} found",
-  "Not found": "Not found",
   "404: Not Found": "404: Not Found",
   "{{labels}} content is not available in the catalog at this time due to loading failures.": "{{labels}} content is not available in the catalog at this time due to loading failures.",
   "Timed out fetching new data. The data below is stale.": "Timed out fetching new data. The data below is stale.",
